Remove some conditionals from Interactions
PHYS-16 #time 5m

Technically this means I'll be doing a full vector operation each
iteration, rather than an if statement everytime and a vector operation
only when a constant gravity field is active. Now, if there is no field,
just add the zero vector to each object.

I think this is worthwhile for simplicity.