Still trying to debug merging code
PHYS-16 #time 20m

This seems to be working fine in the nonrandom code, but I'm still
getting the weird behavior where both objects in a merge end up getting
eliminated when I try to run the generic bodyFormation code.

It seems that all the contact happens in pairs due to the way that I
"randomly" place objects. Deleting 2 objects at the end of the loops
might be causing the problem.